requirements.txt
alabaster==0.7.12
appdirs==1.4.4
Babel==2.9.1
backcall==0.2.0
bagpy==0.4.5
bitstring==3.1.7
catkin-pkg==0.4.23
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
decorator==5.0.9
distro==1.5.0
docutils==0.16
future==0.18.2
gnupg==2.3.1
idna==2.10
imagesize==1.2.0
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
livereload==2.6.3
lunr==0.5.8
m2r2==0.2.7
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.2.1
matplotlib-inline==0.1.2
mistune==0.8.4
mkdocs==1.1.2
nltk==3.6.2
numpy==1.20.3
packaging==20.9
pandas==1.2.4
parso==0.8.2
pathlib==1.0.1
pickleshare==0.7.5
prompt-toolkit==3.0.18
psutil==5.8.0
py3rosmsgs==1.18.1
pycryptodomex==3.10.1
pydash==5.0.0
Pygments==2.9.0
pymdown-extensions==8.2
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
recommonmark==0.7.1
regex==2021.4.4
requests==2.25.1
rinoh-typeface-dejavuserif==0.1.3
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.5.2
rospkg==1.3.0
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
snowballstemmer==2.1.0
Sphinx==3.2.1
sphinx-autodoc-typehints==1.4.0
sphinx-bootstrap-theme==0.7.1
sphinx-markdown-parser==0.2.4
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
unify==0.5
untokenize==0.1.1
urllib3==1.26.5
wcwidth==0.2.5
yapf==0.31.0
ytsphinx==1.2.1.dev20200430
